make stat trains doubles

+++ src/fight.c
@@ -2154,15 +2154,15 @@ violence_update(void)
 		}
 		/* Gravity area/room effects and bonus combat stats, weighted clothing */
 		if (!IS_NPC(ch)) {
-			double totalrgrav = 0;
-			double addedrweight = 0;
-			double playerrweight = 0;
-			int	weightstatmult = 0;
-			int	weightstat = 0;
-			int	weighttrainmult = 0;
-			int	weighttrain = 0;
-			int	damrange = 0;
-			int	gravdam = 0;
+			double 	totalrgrav = 0;
+			double 	addedrweight = 0;
+			double 	playerrweight = 0;
+			double	weightstatmult = 0;
+			double	weightstat = 0;
+			double	weighttrainmult = 0;
+			double	weighttrain = 0;
+			int		damrange = 0;
+			int		gravdam = 0;
 			
 			
 			damrange = number_range(1, 3);
@@ -2170,10 +2170,10 @@ violence_update(void)
 			addedrweight = (double) weightedtraining(ch) / 100000;
 			playerrweight = (double) 1 + addedrweight;
 			totalrgrav = (double) ch->gravSetting * playerrweight;
-			weightstatmult = (totalrgrav / 50) + 1;
-			weightstat = 5 * weightstatmult;
-			weighttrainmult = (totalrgrav / 50) + 1;
-			weighttrain = 5 * weighttrainmult;
+			weightstatmult = (double) ((double)totalrgrav / 50) + 1;
+			weightstat = (double) 5 * weightstatmult;
+			weighttrainmult = (double) ((double)totalrgrav / 50) + 1;
+			weighttrain = (double) 5 * weighttrainmult;
 			
 			if (totalrgrav > 1) {
 				gravdam = damrange * totalrgrav;
